+++ inc/common.class.php
@@ -806,7 +806,7 @@ JAVASCRIPT;
          $newMenu['faq'] = $menus['faq'];
          $newMenu['faq']['default'] = Plugin::getWebDir('formcreator', false) . '/front/knowbaseitem.php';
       }
-      if (Session::haveRight("reservation", ReservationItem::RESERVEANITEM)) {
+      if (Session::haveRightsOr("reservation", [READ, ReservationItem::RESERVEANITEM])) {
          if (isset($menus['reservation'])) {
             $newMenu['reservation'] = $menus['reservation'];
          }
@@ -871,7 +871,7 @@ JAVASCRIPT;
 
       if (PluginFormcreatorEntityconfig::getUsedConfig('is_dashboard_visible', Session::getActiveEntity()) == PluginFormcreatorEntityconfig::CONFIG_DASHBOARD_VISIBLE) {
          if (version_compare(GLPI_VERSION, '10.0.3') > 0) {
-            $dashboard = new Glpi\Dashboard\Grid('plugin_formcreator_issue_counters', 33, 1, 'mini_core');
+            $dashboard = new Glpi\Dashboard\Grid('plugin_formcreator_issue_counters', 33, 2, 'mini_core');
          } else {
             $dashboard = new Glpi\Dashboard\Grid('plugin_formcreator_issue_counters', 33, 0, 'mini_core');
          }