+++ inc/field/datetimefield.class.php
@@ -46,6 +46,7 @@ class DatetimeField extends PluginFormcreatorAbstractField
    protected $fields = null;
 
    const DATE_FORMAT = 'Y-m-d H:i:s';
+   const DATE_ZERO   = '0000-00-00 00:00:00';
 
    public function isPrerequisites(): bool {
       return true;
@@ -131,7 +132,7 @@ class DatetimeField extends PluginFormcreatorAbstractField
          return true;
       }
 
-      $check = DateTime::createFromFormat(self::DATE_FORMAT, $value);
+      $check = $this->getDateFromValue($value);
       return $check !== false;
    }
 
@@ -143,14 +144,25 @@ class DatetimeField extends PluginFormcreatorAbstractField
       return true;
    }
 
-   public function equals($value): bool {
-      if ($this->value === '') {
-         $answer = '0000-00-00 00:00:00';
-      } else {
-         $answer = $this->value;
+   /**
+    * Convert a string value into DateTime object
+    *
+    * @param string $value
+    * @return false|DateTime
+    */
+   protected function getDateFromValue(string $value) {
+      if (empty($value)) {
+         $value = self::DATE_ZERO;
       }
-      $answerDatetime = DateTime::createFromFormat(self::DATE_FORMAT, $answer);
-      $compareDatetime = DateTime::createFromFormat(self::DATE_FORMAT, $value);
+      return DateTime::createFromFormat(self::DATE_FORMAT, $value);
+   }
+
+   public function equals($value): bool {
+      $answerDatetime = $this->getDateFromValue($this->value ?? '');
+      $compareDatetime = $this->getDateFromValue($value);
+      if ($compareDatetime === false) {
+         return false;
+      }
       return $answerDatetime == $compareDatetime;
    }
 
@@ -159,13 +171,11 @@ class DatetimeField extends PluginFormcreatorAbstractField
    }
 
    public function greaterThan($value): bool {
-      if (empty($this->value)) {
-         $answer = '0000-00-00 00:00:00';
-      } else {
-         $answer = $this->value;
+      $answerDatetime = $this->getDateFromValue($this->value ?? '');
+      $compareDatetime = $this->getDateFromValue($value);
+      if ($compareDatetime === false) {
+         return true;
       }
-      $answerDatetime = DateTime::createFromFormat(self::DATE_FORMAT, $answer);
-      $compareDatetime = DateTime::createFromFormat(self::DATE_FORMAT, $value);
       return $answerDatetime > $compareDatetime;
    }
 
@@ -187,7 +197,7 @@ class DatetimeField extends PluginFormcreatorAbstractField
       }
 
       if ($input[$key] != ''
-         && DateTime::createFromFormat(self::DATE_FORMAT, $input[$key]) === false
+         && $this->getDateFromValue($input[$key]) === false
       ) {
          return false;
       }