+++ inc/field/dropdownfield.class.php
@@ -41,6 +41,7 @@ use Session;
 use DBUtils;
 use Document;
 use Dropdown;
+use CommonDBTM;
 use CommonITILActor;
 use CommonITILObject;
 use CommonTreeDropdown;
@@ -57,6 +58,7 @@ use Search;
 use SLA;
 use SLM;
 use OLA;
+use QueryExpression;
 use QuerySubQuery;
 use QueryUnion;
 use GlpiPlugin\Formcreator\Exception\ComparisonException;
@@ -227,9 +229,11 @@ class DropdownField extends PluginFormcreatorAbstractField
             $currentUser = Session::getLoginUserID();
             if (!Session::haveRight(Ticket::$rightname, Ticket::READMY) && !Session::haveRight(Ticket::$rightname, Ticket::READGROUP)) {
                // No right to view any ticket, then force the dropdown to be empty
-               $dparams_cond_crit['OR'] = new \QueryExpression('0=1');
+               $dparams_cond_crit['OR'] = new QueryExpression('0=1');
                break;
             }
+            $tickets_filter = ['users_id_recipient' => $currentUser];
+
             if (Session::haveRight(Ticket::$rightname, Ticket::READMY)) {
                $requestersObserversQuery = new QuerySubQuery([
                   'SELECT' => 'tickets_id',
@@ -239,34 +243,23 @@ class DropdownField extends PluginFormcreatorAbstractField
                      'type' => [CommonITILActor::REQUESTER, CommonITILActor::OBSERVER]
                   ],
                ]);
-               $dparams_cond_crit['OR'] = [
+               $tickets_filter[] = [
                   'id' => $requestersObserversQuery,
-                  'users_id_recipient' => $currentUser,
                ];
             }
-            if (Session::haveRight(Ticket::$rightname, Ticket::READGROUP)) {
-               $sub_query = [
+
+            if (Session::haveRight(Ticket::$rightname, Ticket::READGROUP) && count($_SESSION['glpigroups']) > '0') {
+               $requestersObserversGroupsQuery = new QuerySubQuery([
                   'SELECT' => 'tickets_id',
                   'FROM' => Group_Ticket::getTable(),
                   'WHERE' => [
-                     'type' => [CommonITILActor::REQUESTER, CommonITILActor::OBSERVER]
+                     'type' => [CommonITILActor::REQUESTER, CommonITILActor::OBSERVER],
+                     'groups_id' => $_SESSION['glpigroups'],
                   ],
-               ];
-               if (count($_SESSION['glpigroups']) > '0') {
-                  $sub_query['WHERE']['groups_id'] = $_SESSION['glpigroups'];
-               }
-               $requestersObserversGroupsQuery = new QuerySubQuery($sub_query);
-               if (!isset($dparams_cond_crit['OR']['id'])) {
-                  $dparams_cond_crit['OR'] = [
-                     'id' => $requestersObserversGroupsQuery,
-                  ];
-               } else {
-                  $dparams_cond_crit['OR']['id'] = new QueryUnion([
-                     $dparams_cond_crit['OR']['id'],
-                     $requestersObserversGroupsQuery,
-                  ]);
-               }
+               ]);
+               $tickets_filter[] = ['id' => $requestersObserversGroupsQuery];
             }
+            $dparams_cond_crit['OR'] = $tickets_filter;
             break;
 
          default:
@@ -312,14 +305,15 @@ class DropdownField extends PluginFormcreatorAbstractField
       // Set specific root if defined (CommonTreeDropdown)
       $baseLevel = 0;
       if (isset($decodedValues['show_tree_root'])
-         && (int) $decodedValues['show_tree_root'] > 0
+         && ((int) $decodedValues['show_tree_root'] > 0
+           || $itemtype == Entity::class && (int) $decodedValues['show_tree_root'] > -1)
       ) {
          $sons = (new DBUtils)->getSonsOf(
             $itemtype::getTable(),
             $decodedValues['show_tree_root']
          );
          $decodedValues['selectable_tree_root'] = $decodedValues['selectable_tree_root'] ?? '1';
-         if (!isset($decodedValues['selectable_tree_root']) || $decodedValues['selectable_tree_root'] == '0') {
+         if ($decodedValues['selectable_tree_root'] == '0') {
             unset($sons[$decodedValues['show_tree_root']]);
          }
 
@@ -708,26 +702,17 @@ class DropdownField extends PluginFormcreatorAbstractField
       // $questionID = $question->fields['id'];
       $questionID = $this->getQuestion()->getID();
 
-      // We need english locale to search searchOptions by name
-      $oldLocale = $TRANSLATE->getLocale();
-      $TRANSLATE->setLocale("en_GB");
-
       // Load target item from DB
-      // $itemtype = $question->getField('values');
       $itemtype = $this->question->fields['itemtype'];
-
-      // Itemtype is stored in plaintext for GlpiselectField and in
-      // json for DropdownField
-      $json = json_decode($itemtype);
-
-      if ($json) {
-         $itemtype = $json->itemtype;
-      }
 
       // Safe check
       if (empty($itemtype) || !class_exists($itemtype)) {
          return $content;
       }
+
+      // We need english locale to search searchOptions by name
+      $oldLocale = $TRANSLATE->getLocale();
+      $TRANSLATE->setLocale("en_GB");
 
       $item = new $itemtype;
       $item->getFromDB($answer);
@@ -743,6 +728,11 @@ class DropdownField extends PluginFormcreatorAbstractField
          // Convert Property_Name to Property Name
          $property = str_replace("_", " ", $property);
          $searchOption = $item->getSearchOptionByField("name", $property);
+         if (count($searchOption) == 0) {
+            trigger_error("No search option found for $property", E_USER_WARNING);
+            $TRANSLATE->setLocale($oldLocale);
+            return $content;
+         }
 
          // Execute search
          $data = Search::prepareDatasForSearch(get_class($item), [