+++ inc/section.class.php
@@ -239,6 +239,20 @@ PluginFormcreatorTranslatableInterface
       ) + 1;
       $newSectionId = static::import($linker, $export, $this->fields[$formFk]);
 
+      // Before importing the section, we need to give to the linker the questions
+      // used in the conditions of the section being duplicated
+      $conditions = (new PluginFormcreatorCondition())->find([
+         'itemtype' => self::getType(),
+         'items_id' => $this->getID()
+      ]);
+      foreach ($conditions as $row) {
+         $question = PluginFormcreatorQuestion::getById($row['plugin_formcreator_questions_id']);
+         if ($question === null || $question === false) {
+            continue;
+         }
+         $linker->addObject($row['plugin_formcreator_questions_id'], $question);
+      }
+
       if ($newSectionId === false) {
          return false;
       }
@@ -552,7 +566,7 @@ PluginFormcreatorTranslatableInterface
       ]);
       $formId = $this->fields[PluginFormcreatorForm::getForeignKeyField()];
       $onclick = 'plugin_formcreator.showSectionForm(' . $formId . ', ' . $sectionId . ');';
-      $html = '<a href="#" onclick=' . $onclick . '" data-field="name">';
+      $html = '<a href="#" onclick="' . $onclick . '" data-field="name">';
       $html .= "<sup class='plugin_formcreator_conditions_count' title='" . __('Count of conditions', 'formcreator') ."'>$nb</sup>";
       $html .= '<span>';
       $html .= empty($this->fields['name']) ? '(' . $sectionId . ')' : $this->fields['name'];